make install: use symlink for installed cowthink.1 man page
Use a symlink for the installed cowthink.1 man page.

As we are already using a symlink for the cowthink executable,
we can reasonably assume the system supports symlinks.

This means the cowthink.1 file from the source repository is really
only used for when people run "man ./cowthink.1" in a cowsay source
tree during development.